[feature-102/adding_room_orm] Minor refactor on imports
