+++ app/src/main/java/com/manday/marvel/data/datasource/db/room/CharacterDao.kt
@@ -1,0 +1,20 @@
+package com.manday.marvel.data.datasource.db.room
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.Query
+import com.manday.marvel.data.datasource.db.models.MarvelCharacter
+
+@Dao
+interface CharacterDao {
+
+    @Query("SELECT * FROM marvelcharacter")
+    fun getAll(): List<MarvelCharacter>
+
+    @Insert
+    fun insertAll(characters: List<MarvelCharacter>)
+
+    @Delete
+    fun delete(character: MarvelCharacter)
+}
