+++ app/src/main/java/com/manday/marvel/domain/repository/CharacterRepository.kt
@@ -1,6 +1,6 @@
 package com.manday.marvel.domain.repository
 
-import com.manday.marvel.data.models.CharacterEntity
+import com.manday.marvel.data.datasource.net.models.CharacterEntity
 import kotlinx.coroutines.flow.Flow
 
 interface CharacterRepository {
